refactor: merge navbar components for consistency
- Integrated GenerationModal from MinimalHeader into Navbar
- Added conditional rendering (Generate button on homepage, Settings elsewhere)
- Preserved event listeners for review questions and keyboard shortcuts
- Simplified ConditionalNavbar logic to use unified Navbar component
- Deleted obsolete MinimalHeader component
- Fixed navbar height consistency with h-16 class (64px)
